Just moving service interfaces to a new folder, remove some unusing imports and put some regions
