+++ PontoFacil/PontoFacil/PontoFacil.csproj
@@ -105,11 +105,11 @@
     <Compile Include="Repositories\IRepository.cs" />
     <Compile Include="Services\ClockInService.cs" />
     <Compile Include="Services\HistoryService.cs" />
-    <Compile Include="Services\IClockInService.cs" />
-    <Compile Include="Services\IHistoryService.cs" />
-    <Compile Include="Services\IPersistencyService.cs" />
-    <Compile Include="Services\IPlanningService.cs" />
-    <Compile Include="Services\ISettingsService.cs" />
+    <Compile Include="Services\Interfaces\IClockInService.cs" />
+    <Compile Include="Services\Interfaces\IHistoryService.cs" />
+    <Compile Include="Services\Interfaces\IPersistencyService.cs" />
+    <Compile Include="Services\Interfaces\IPlanningService.cs" />
+    <Compile Include="Services\Interfaces\ISettingsService.cs" />
     <Compile Include="Services\NotificationService.cs" />
     <Compile Include="Services\PersistencyService.cs" />
     <Compile Include="Services\PlanningService.cs" />
