+++ PontoFacil/PontoFacil/Services/Interfaces/IClockInService.cs
@@ -1,12 +1,11 @@
 ﻿using PontoFacil.Models;
 using System;
 
-namespace PontoFacil.Services
+namespace PontoFacil.Services.Interfaces
 {
     public interface IClockInService
     {
         void Register(DateTime date);
-
         ClockIn getClockInById(DateTime datetime);
     }
 }
