+++ PontoFacil/PontoFacil/Services/SettingsService.cs
@@ -1,4 +1,5 @@
 ﻿using PontoFacil.Models;
+using PontoFacil.Services.Interfaces;
 
 namespace PontoFacil.Services
 {
