+++ PontoFacil/PontoFacil/ViewModels/HistoryPageViewModel.cs
@@ -1,15 +1,10 @@
 ﻿using PontoFacil.Models;
-using PontoFacil.Services;
+using PontoFacil.Services.Interfaces;
 using Prism.Commands;
 using Prism.Windows.Mvvm;
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Windows.UI.Popups;
-using Windows.UI.Xaml;
 
 namespace PontoFacil.ViewModels
 {
