+++ PontoFacil/PontoFacil/ViewModels/PlanningPageViewModel.cs
@@ -1,8 +1,7 @@
 ﻿using PontoFacil.Models;
-using PontoFacil.Services;
+using PontoFacil.Services.Interfaces;
 using Prism.Commands;
 using Prism.Windows.Mvvm;
-using System;
 
 namespace PontoFacil.ViewModels
 {
